ai-tools/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
pillow
requests
redis
rq
psycopg2-binary 
replicate 
requests
pydantic


